+++ package.json
@@ -20,15 +20,16 @@
     "dependencies": {
         "@next/bundle-analyzer": "^12.1.6",
         "esbuild": "^0.14.39",
-        "fast-memoize": "^2.5.2",
-        "gray-matter": "^4.0.3",
         "mdx-bundler": "^9.0.1",
+        "gray-matter": "^4.0.3",
+        "jotai": "^1.7.0",
+        "fast-memoize": "^2.5.2",
         "next": "12.1.6",
         "next-compose-plugins": "^2.2.1",
         "next-seo": "^5.4.0",
-        "react": "^18.1.0",
-        "react-dom": "^18.1.0",
-        "recoil": "^0.6.1",
+        "preact": "^10.7.2",
+        "react": "^17.0.2",
+        "react-dom": "^17.0.2",
         "rehype-katex": "^6.0.2",
         "rehype-prism-plus": "^1.3.2",
         "remark-gfm": "^3.0.1",
@@ -41,13 +42,12 @@
         "@types/mdx": "^2.0.2",
         "@types/mdx-js__react": "^1.5.5",
         "@types/node": "17.0.35",
-        "@types/react": "18.0.9",
+        "@types/react": "^17.0.2",
         "@types/styled-components": "^5.1.25",
         "eslint": "^8.16.0",
         "eslint-config-next": "12.1.6",
         "ts-node": "^10.8.0",
         "tsconfig-paths": "^4.0.0",
         "typescript": "4.7.2"
-    },
-    "packageManager": "yarn@3.2.1"
+    }
 }
